Added code comments.

+++ Software/ajb_microrf51_manual/io_ctrl.ino
@@ -9,6 +9,7 @@ volatile uint8_t pwm_counter;
 uint8_t pwm_comp0;
 uint8_t pwm_comp1;
 
+// Soft PWM (temporary)
 void pwm_runner() {
   pwm_counter++;
 
@@ -23,16 +24,16 @@ void pwm_runner() {
     NRF_GPIO->OUTCLR  = (1 << R_MOTOR_PWM);
 }
 
-void motor_setSpeedLR(int32_t Lspeed, int32_t Rspeed) {
+void motor_setSpeedLR(int16_t Lspeed, int16_t Rspeed) {
   if ((Lspeed > 255) || (Lspeed < -255))
     return;
   if ((Rspeed > 255) || (Rspeed < -255))
     return;
     
   // Left motor
-  
   digitalWrite(L_MOTOR_DIR, Lspeed >= 0 ? L_FORWARD : L_BACKWARD);
   pwm_comp0 = (Lspeed >= 0 ? Lspeed : -Lspeed);
+  
   // Right motor
   digitalWrite(R_MOTOR_DIR, Rspeed >= 0 ? R_FORWARD : R_BACKWARD);
   pwm_comp1 = (Rspeed >= 0 ? Rspeed : -Rspeed);
